+++ client/src/components/reviews/Review.js
@@ -13,7 +13,10 @@ export const Review = ({ review, onDelete }) => {
     const navigate = useNavigate()
 
     const handleDelete = () => {
-        onDelete(review.id)
+        if (window.confirm("Are you sure you want to remove this review?")) {
+            window.alert("Review removed from the ether.")
+            onDelete(review.id)
+        }
     }
 
     return (
